`pre-commit autoupdate`

+++ tests/test_ops.py
@@ -9,12 +9,15 @@ arr2d = np.array([[1, 2, 0, 0], [0, 1, 0, 1]])
 
 arr1d = np.array([1, 1, 2, 3])
 
-parametrize_optimizer = pytest.mark.parametrize("opt", [finch.DefaultScheduler(), finch.GalleyScheduler()])  
+parametrize_optimizer = pytest.mark.parametrize(
+    "opt", [finch.DefaultScheduler(), finch.GalleyScheduler()]
+)
+
 
 @parametrize_optimizer
 def test_eager(arr3d, opt):
     finch.set_optimizer(opt)
-    
+
     A_finch = finch.Tensor(arr3d)
     B_finch = finch.Tensor(arr2d)
 
@@ -26,7 +29,7 @@ def test_eager(arr3d, opt):
 @parametrize_optimizer
 def test_lazy_mode(arr3d, opt):
     finch.set_optimizer(opt)
-    
+
     A_finch = finch.Tensor(arr3d)
     B_finch = finch.Tensor(arr2d)
     C_finch = finch.Tensor(arr1d)
@@ -76,7 +79,7 @@ def test_lazy_mode(arr3d, opt):
 @parametrize_optimizer
 def test_elemwise_ops_1_arg(arr3d, func_name, opt):
     finch.set_optimizer(opt)
-    
+
     arr = arr3d + 1.6
     A_finch = finch.Tensor(arr)
 
@@ -86,15 +89,15 @@ def test_elemwise_ops_1_arg(arr3d, func_name, opt):
     assert_allclose(actual.todense(), expected)
 
 
-@pytest.mark.parametrize(
-    "func_name", ["real", "imag", "conj"]
-)
+@pytest.mark.parametrize("func_name", ["real", "imag", "conj"])
 @pytest.mark.parametrize("dtype", [np.complex128, np.complex64, np.float64, np.int64])
 @parametrize_optimizer
 def test_elemwise_complex_ops_1_arg(func_name, dtype, opt):
     finch.set_optimizer(opt)
-    
-    arr = np.asarray([[1+1j, 2+2j], [3+3j, 4-4j], [-5-5j, -6-6j]]).astype(dtype)
+
+    arr = np.asarray([[1 + 1j, 2 + 2j], [3 + 3j, 4 - 4j], [-5 - 5j, -6 - 6j]]).astype(
+        dtype
+    )
     arr_finch = finch.asarray(arr)
 
     actual = getattr(finch, func_name)(arr_finch)
@@ -111,7 +114,7 @@ def test_elemwise_complex_ops_1_arg(func_name, dtype, opt):
 @parametrize_optimizer
 def test_elemwise_tensor_ops_1_arg(arr3d, meth_name, opt):
     finch.set_optimizer(opt)
-    
+
     A_finch = finch.Tensor(arr3d)
 
     actual = getattr(A_finch, meth_name)()
@@ -127,7 +130,7 @@ def test_elemwise_tensor_ops_1_arg(arr3d, meth_name, opt):
 @parametrize_optimizer
 def test_elemwise_ops_2_args(arr3d, func_name, opt):
     finch.set_optimizer(opt)
-    
+
     arr2d = np.array([[0, 3, 2, 0], [0, 0, 3, 2]])
     if func_name.startswith("logical"):
         arr3d = arr3d.astype(bool)
@@ -167,7 +170,7 @@ def test_elemwise_ops_2_args(arr3d, func_name, opt):
 @parametrize_optimizer
 def test_elemwise_tensor_ops_2_args(arr3d, meth_name, opt):
     finch.set_optimizer(opt)
-    
+
     arr2d = np.array([[2, 3, 2, 3], [3, 2, 3, 2]])
     A_finch = finch.Tensor(arr3d)
     B_finch = finch.Tensor(arr2d)
@@ -183,7 +186,7 @@ def test_elemwise_tensor_ops_2_args(arr3d, meth_name, opt):
 @parametrize_optimizer
 def test_reductions(arr3d, func_name, axis, opt):
     finch.set_optimizer(opt)
-    
+
     A_finch = finch.Tensor(arr3d)
 
     actual = getattr(finch, func_name)(A_finch, axis=axis)
@@ -205,9 +208,11 @@ def test_reductions(arr3d, func_name, axis, opt):
     ],
 )
 @parametrize_optimizer
-def test_sum_prod_dtype_arg(arr3d, func_name, axis, in_dtype, dtype, expected_dtype, opt):
+def test_sum_prod_dtype_arg(
+    arr3d, func_name, axis, in_dtype, dtype, expected_dtype, opt
+):
     finch.set_optimizer(opt)
-    
+
     arr_finch = finch.asarray(np.abs(arr3d), dtype=in_dtype)
 
     actual = getattr(finch, func_name)(arr_finch, axis=axis, dtype=dtype).todense()
@@ -236,7 +241,7 @@ def test_sum_prod_dtype_arg(arr3d, func_name, axis, in_dtype, dtype, expected_dt
 @parametrize_optimizer
 def test_tensordot(arr3d, storage, opt):
     finch.set_optimizer(opt)
-    
+
     A_finch = finch.Tensor(arr1d)
     B_finch = finch.Tensor(arr2d)
     C_finch = finch.Tensor(arr3d)
@@ -267,7 +272,7 @@ def test_tensordot(arr3d, storage, opt):
 @parametrize_optimizer
 def test_matmul(arr2d, arr3d, opt):
     finch.set_optimizer(opt)
-    
+
     A_finch = finch.Tensor(arr2d)
     B_finch = finch.Tensor(arr2d.T)
     C_finch = finch.permute_dims(A_finch, (1, 0))
@@ -287,7 +292,7 @@ def test_matmul(arr2d, arr3d, opt):
 @parametrize_optimizer
 def test_negative__mod__(opt):
     finch.set_optimizer(opt)
-    
+
     arr = np.array([-1, 0, 0, -2, -3, 0])
     arr_finch = finch.asarray(arr)
